moved build_simple_transaction_empty_gas_payment from app client to standalone
