+++ tests/application_client/sui.py
@@ -210,77 +210,6 @@ class SuiClient:
 
         return [tx, object_list]
 
-    def build_simple_transaction_empty_gas_payment(
-        self, sender_addr: str, destination: str, send_amount: int, fees: int
-    ) -> [bytes, [bytes]]:
-        """
-        Build a TransferToken transaction with empty gas_data.payment (SIP-58).
-        Gas is paid from address balance. Uses Input(0) for the coin, not GasCoin.
-        """
-        tx = b''
-        gas_budget = fees
-
-        intent_bsc = Intent.encode(Intent.from_list([0, 0, 0]))
-        tx += intent_bsc
-
-        amount_bytes = list(send_amount.to_bytes(8, byteorder='little'))
-        recepient_addr = list(bytes.fromhex(destination[2:]))
-
-        obj_info = USDC_OBJECTS_BY_AMOUNT[send_amount]
-        object_list = [base64.b64decode(obj_info['obj'])]
-
-        tx_data_v1 = TransactionDataV1(
-            TransactionKind=TransactionKind(
-                "ProgrammableTransaction",
-                ProgrammableTransaction(
-                    Inputs=[
-                        CallArg(
-                            "Object",
-                            ObjectArg(
-                                "ImmOrOwnedObject",
-                                ObjectReference(
-                                    ObjectID=Address.from_str(obj_info['object_id']),
-                                    SequenceNumber=obj_info['version'],
-                                    ObjectDigest=Digest.from_str(obj_info['digest']),
-                                ),
-                            ),
-                        ),
-                        CallArg("Pure", amount_bytes),
-                        CallArg("Pure", recepient_addr),
-                    ],
-                    Command=[
-                        Command(
-                            "SplitCoin",
-                            SplitCoin(
-                                FromCoin=Argument("Input", 0),
-                                Amount=[Argument("Input", 1)],
-                            ),
-                        ),
-                        Command(
-                            "TransferObjects",
-                            TransferObjects(
-                                Objects=[Argument("Result", 0)],
-                                Address=Argument("Input", 2),
-                            ),
-                        ),
-                    ],
-                ),
-            ),
-            Sender=Address.from_str(sender_addr),
-            GasData=GasData(
-                Payment=[],  # SIP-58: empty payment, gas from address balance
-                Owner=Address.from_str(sender_addr),
-                Price=1,
-                Budget=gas_budget,
-            ),
-            TransactionExpiration=TransactionExpiration("None"),
-        )
-
-        tx_data = TransactionData("V1", tx_data_v1)
-        tx += TransactionData.encode(tx_data)
-
-        return [tx, object_list]
-
     def sign_transaction(self, path: bytes, txn: Union[str, bytes, bytearray], object_list: Optional[list[bytes]] = None) -> bytes:
         """
         Sign a transaction with the key at a BIP32 path.
